Split For.hpp into For.hpp and ForICount.hpp
