+++ include/RAJA/pattern/kernel/ForICount.hpp
@@ -25,8 +25,8 @@
 //~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~//
 
 
-#ifndef RAJA_pattern_kernel_For_HPP
-#define RAJA_pattern_kernel_For_HPP
+#ifndef RAJA_pattern_kernel_ForICount_HPP
+#define RAJA_pattern_kernel_ForICount_HPP
 
 #include "RAJA/config.hpp"
 
@@ -45,19 +45,23 @@ namespace statement
 /*!
  * A RAJA::kernel statement that implements a single loop.
  * Assigns the loop iterate to argument ArgumentId
+ * Assigns the loop index to param ParamId
  *
  */
 template <camp::idx_t ArgumentId,
+          typename ParamId,
           typename ExecPolicy = camp::nil,
           typename... EnclosedStmts>
-struct For : public internal::ForList,
+struct ForICount : public internal::ForList,
              public internal::ForTraitBase<ArgumentId, ExecPolicy>,
              public internal::Statement<ExecPolicy, EnclosedStmts...> {
 
+  static_assert(std::is_base_of<internal::ParamBase, ParamId>::value,
+                "Inappropriate ParamId, ParamId must be of type "
+                "RAJA::Statement::Param< # >");
   // TODO: add static_assert for valid policy in Pol
   using execution_policy_t = ExecPolicy;
 };
-
 
 }  // end namespace statement
 
@@ -65,36 +69,39 @@ namespace internal
 {
 
 /*!
- * A generic RAJA::kernel forall_impl loop wrapper for statement::For
+ * A generic RAJA::kernel forall_impl loop wrapper for statement::ForICount
  * Assigns the loop index to offset ArgumentId
- *
+ * Assigns the loop index to param ParamId
  */
-template <camp::idx_t ArgumentId, typename Data, typename... EnclosedStmts>
-struct ForWrapper : public GenericWrapper<Data, EnclosedStmts...> {
+template <camp::idx_t ArgumentId, typename ParamId, typename Data,
+          typename... EnclosedStmts>
+struct ForICountWrapper : public GenericWrapper<Data, EnclosedStmts...> {
 
   using Base = GenericWrapper<Data, EnclosedStmts...>;
   using Base::Base;
-  using privatizer = NestedPrivatizer<ForWrapper>;
+  using privatizer = NestedPrivatizer<ForICountWrapper>;
 
   template <typename InIndexType>
   RAJA_INLINE void operator()(InIndexType i)
   {
     Base::data.template assign_offset<ArgumentId>(i);
+    Base::data.template assign_param<ParamId>(i);
     Base::exec();
   }
 };
 
 
 /*!
- * A generic RAJA::kernel forall_impl executor for statement::For
+ * A generic RAJA::kernel forall_impl executor for statement::ForICount
  *
  *
  */
 template <camp::idx_t ArgumentId,
+          typename ParamId,
           typename ExecPolicy,
           typename... EnclosedStmts>
 struct StatementExecutor<
-    statement::For<ArgumentId, ExecPolicy, EnclosedStmts...>> {
+    statement::ForICount<ArgumentId, ParamId, ExecPolicy, EnclosedStmts...>> {
 
 
   template <typename Data>
@@ -102,7 +109,8 @@ struct StatementExecutor<
   {
 
     // Create a wrapper, just in case forall_impl needs to thread_privatize
-    ForWrapper<ArgumentId, Data, EnclosedStmts...> for_wrapper(data);
+    ForICountWrapper<ArgumentId, ParamId, Data,
+                     EnclosedStmts...> for_wrapper(data);
 
     auto len = segment_length<ArgumentId>(data);
     using len_t = decltype(len);
