+++ include/RAJA/policy/simd/kernel/For.hpp
@@ -172,46 +172,6 @@ struct StatementExecutor<
   }
 };
 
-/*!
- * RAJA::kernel forall_impl executor specialization for statement::ForICount.
- * Assumptions: RAJA::simd_exec is the inner most policy,
- * only one lambda is used, no reductions are done within the lambda.
- * Assigns the loop index to offset ArgumentId
- * Assigns the loop index to param ParamId
- */
-template <camp::idx_t ArgumentId, typename ParamId,
-          typename... EnclosedStmts>
-struct StatementExecutor<
-    statement::ForICount<ArgumentId, ParamId, RAJA::simd_exec,
-                         EnclosedStmts...>> {
-
-  template <typename Data>
-  static RAJA_INLINE void exec(Data &&data)
-  {
-
-    auto iter = get<ArgumentId>(data.segment_tuple);
-    auto begin = std::begin(iter);
-    auto end = std::end(iter);
-    auto distance = std::distance(begin, end);
-
-    RAJA_SIMD
-    for (decltype(distance) i = 0; i < distance; ++i) {
-
-      // Offsets and parameters need to be privatized
-      auto offsets = data.offset_tuple;
-      auto params = data.param_tuple;
-      get<ArgumentId>(offsets) = i;
-      get<ParamId::param_idx>(params) = i;
-
-      Invoke_all_Lambda<0, EnclosedStmts...>::lambda_special(
-          camp::idx_seq_from_t<decltype(offsets)>{},
-          camp::idx_seq_from_t<decltype(params)>{},
-          data,
-          offsets,
-          params);
-    }
-  }
-};
 
 }  // namespace internal
 }  // end namespace RAJA
